requirements.txt
google-generativeai
pytesseract
python-dotenv